parameter print updated

+++ 2/censored.c
@@ -7,10 +7,10 @@
 
 int main(int argc, char const *argv[]) {
   // print parameters
-  for (size_t argarg = 0; argarg < argc; argarg++) {
+  for (int argarg = 0; argarg < argc; argarg++) {
     printf("%s\n", argv[argarg]);
   }
-  
+
   // calc argument lengths
   unsigned int argLengths[argc];
   for (unsigned int argLen_i; argLen_i < argc; argLen_i++) {
